Extract core data managed models and helpers into separate files
